Removed temp files. Changed event and main app to match temp files. Commented out old code for review
